requirements.txt
FPDF2
dotenv
inspect
requests
tabulate
bs4 import BeautifulSoup
html.parser import HTMLParser
